bter: drop in favour of gateio
