--- js/bter.js
@@ -1,298 +1,0 @@
-'use strict';
-
-//  ---------------------------------------------------------------------------
-
-const Exchange = require ('./base/Exchange');
-const { ExchangeError } = require ('./base/errors');
-
-//  ---------------------------------------------------------------------------
-
-module.exports = class bter extends Exchange {
-    describe () {
-        return this.deepExtend (super.describe (), {
-            'id': 'bter',
-            'name': 'Bter',
-            'countries': [ 'VG', 'CN' ], // British Virgin Islands, China
-            'version': '2',
-            'has': {
-                'CORS': false,
-                'createMarketOrder': false,
-                'fetchTickers': true,
-                'withdraw': true,
-            },
-            'urls': {
-                'logo': 'https://user-images.githubusercontent.com/1294454/27980479-cfa3188c-6387-11e7-8191-93fc4184ba5c.jpg',
-                'api': {
-                    'public': 'https://data.bter.com/api',
-                    'private': 'https://api.bter.com/api',
-                },
-                'www': 'https://bter.com',
-                'doc': 'https://bter.com/api2',
-                'fees': 'https://bter.com/fee',
-            },
-            'api': {
-                'public': {
-                    'get': [
-                        'pairs',
-                        'marketinfo',
-                        'marketlist',
-                        'tickers',
-                        'ticker/{id}',
-                        'orderBook/{id}',
-                        'trade/{id}',
-                        'tradeHistory/{id}',
-                        'tradeHistory/{id}/{tid}',
-                    ],
-                },
-                'private': {
-                    'post': [
-                        'balances',
-                        'depositAddress',
-                        'newAddress',
-                        'depositsWithdrawals',
-                        'buy',
-                        'sell',
-                        'cancelOrder',
-                        'cancelAllOrders',
-                        'getOrder',
-                        'openOrders',
-                        'tradeHistory',
-                        'withdraw',
-                    ],
-                },
-            },
-        });
-    }
-
-    async fetchMarkets () {
-        let response = await this.publicGetMarketinfo ();
-        let markets = this.safeValue (response, 'pairs');
-        if (!markets)
-            throw new ExchangeError (this.id + ' fetchMarkets got an unrecognized response');
-        let result = [];
-        for (let i = 0; i < markets.length; i++) {
-            let market = markets[i];
-            let keys = Object.keys (market);
-            let id = keys[0];
-            let details = market[id];
-            let [ base, quote ] = id.split ('_');
-            base = base.toUpperCase ();
-            quote = quote.toUpperCase ();
-            base = this.commonCurrencyCode (base);
-            quote = this.commonCurrencyCode (quote);
-            let symbol = base + '/' + quote;
-            let precision = {
-                'amount': details['decimal_places'],
-                'price': details['decimal_places'],
-            };
-            let amountLimits = {
-                'min': details['min_amount'],
-                'max': undefined,
-            };
-            let priceLimits = {
-                'min': undefined,
-                'max': undefined,
-            };
-            let limits = {
-                'amount': amountLimits,
-                'price': priceLimits,
-            };
-            result.push ({
-                'id': id,
-                'symbol': symbol,
-                'base': base,
-                'quote': quote,
-                'info': market,
-                'maker': details['fee'] / 100,
-                'taker': details['fee'] / 100,
-                'precision': precision,
-                'limits': limits,
-            });
-        }
-        return result;
-    }
-
-    async fetchBalance (params = {}) {
-        await this.loadMarkets ();
-        let balance = await this.privatePostBalances ();
-        let result = { 'info': balance };
-        let currencies = Object.keys (this.currencies);
-        for (let i = 0; i < currencies.length; i++) {
-            let currency = currencies[i];
-            let code = this.commonCurrencyCode (currency);
-            let account = this.account ();
-            if ('available' in balance) {
-                if (currency in balance['available']) {
-                    account['free'] = parseFloat (balance['available'][currency]);
-                }
-            }
-            if ('locked' in balance) {
-                if (currency in balance['locked']) {
-                    account['used'] = parseFloat (balance['locked'][currency]);
-                }
-            }
-            account['total'] = this.sum (account['free'], account['used']);
-            result[code] = account;
-        }
-        return this.parseBalance (result);
-    }
-
-    async fetchOrderBook (symbol, params = {}) {
-        await this.loadMarkets ();
-        let orderbook = await this.publicGetOrderBookId (this.extend ({
-            'id': this.marketId (symbol),
-        }, params));
-        let result = this.parseOrderBook (orderbook);
-        result['asks'] = this.sortBy (result['asks'], 0);
-        return result;
-    }
-
-    parseTicker (ticker, market = undefined) {
-        let timestamp = this.milliseconds ();
-        let symbol = undefined;
-        if (market)
-            symbol = market['symbol'];
-        return {
-            'symbol': symbol,
-            'timestamp': timestamp,
-            'datetime': this.iso8601 (timestamp),
-            'high': parseFloat (ticker['high24hr']),
-            'low': parseFloat (ticker['low24hr']),
-            'bid': parseFloat (ticker['highestBid']),
-            'ask': parseFloat (ticker['lowestAsk']),
-            'vwap': undefined,
-            'open': undefined,
-            'close': undefined,
-            'first': undefined,
-            'last': parseFloat (ticker['last']),
-            'change': parseFloat (ticker['percentChange']),
-            'percentage': undefined,
-            'average': undefined,
-            'baseVolume': parseFloat (ticker['quoteVolume']),
-            'quoteVolume': parseFloat (ticker['baseVolume']),
-            'info': ticker,
-        };
-    }
-
-    async fetchTickers (symbols = undefined, params = {}) {
-        await this.loadMarkets ();
-        let tickers = await this.publicGetTickers (params);
-        let result = {};
-        let ids = Object.keys (tickers);
-        for (let i = 0; i < ids.length; i++) {
-            let id = ids[i];
-            let [ baseId, quoteId ] = id.split ('_');
-            let base = baseId.toUpperCase ();
-            let quote = quoteId.toUpperCase ();
-            base = this.commonCurrencyCode (base);
-            quote = this.commonCurrencyCode (quote);
-            let symbol = base + '/' + quote;
-            let ticker = tickers[id];
-            let market = undefined;
-            if (symbol in this.markets)
-                market = this.markets[symbol];
-            if (id in this.markets_by_id)
-                market = this.markets_by_id[id];
-            result[symbol] = this.parseTicker (ticker, market);
-        }
-        return result;
-    }
-
-    async fetchTicker (symbol, params = {}) {
-        await this.loadMarkets ();
-        let market = this.market (symbol);
-        let ticker = await this.publicGetTickerId (this.extend ({
-            'id': market['id'],
-        }, params));
-        return this.parseTicker (ticker, market);
-    }
-
-    parseTrade (trade, market) {
-        let timestamp = this.parse8601 (trade['date']);
-        return {
-            'id': trade['tradeID'],
-            'info': trade,
-            'timestamp': timestamp,
-            'datetime': this.iso8601 (timestamp),
-            'symbol': market['symbol'],
-            'type': undefined,
-            'side': trade['type'],
-            'price': trade['rate'],
-            'amount': this.safeFloat (trade, 'amount'),
-        };
-    }
-
-    async fetchTrades (symbol, since = undefined, limit = undefined, params = {}) {
-        await this.loadMarkets ();
-        let market = this.market (symbol);
-        let response = await this.publicGetTradeHistoryId (this.extend ({
-            'id': market['id'],
-        }, params));
-        return this.parseTrades (response['data'], market, since, limit);
-    }
-
-    async createOrder (symbol, type, side, amount, price = undefined, params = {}) {
-        if (type === 'market')
-            throw new ExchangeError (this.id + ' allows limit orders only');
-        await this.loadMarkets ();
-        let method = 'privatePost' + this.capitalize (side);
-        let order = {
-            'currencyPair': this.marketId (symbol),
-            'rate': price,
-            'amount': amount,
-        };
-        let response = await this[method] (this.extend (order, params));
-        return {
-            'info': response,
-            'id': response['orderNumber'],
-        };
-    }
-
-    async cancelOrder (id, symbol = undefined, params = {}) {
-        await this.loadMarkets ();
-        return await this.privatePostCancelOrder ({ 'orderNumber': id });
-    }
-
-    async withdraw (currency, amount, address, tag = undefined, params = {}) {
-        await this.loadMarkets ();
-        let response = await this.privatePostWithdraw (this.extend ({
-            'currency': currency.toLowerCase (),
-            'amount': amount,
-            'address': address, // Address must exist in you AddressBook in security settings
-        }, params));
-        return {
-            'info': response,
-            'id': undefined,
-        };
-    }
-
-    sign (path, api = 'public', method = 'GET', params = {}, headers = undefined, body = undefined) {
-        let prefix = (api === 'private') ? (api + '/') : '';
-        let url = this.urls['api'][api] + this.version + '/1/' + prefix + this.implodeParams (path, params);
-        let query = this.omit (params, this.extractParams (path));
-        if (api === 'public') {
-            if (Object.keys (query).length)
-                url += '?' + this.urlencode (query);
-        } else {
-            this.checkRequiredCredentials ();
-            let nonce = this.nonce ();
-            let request = { 'nonce': nonce };
-            body = this.urlencode (this.extend (request, query));
-            let signature = this.hmac (this.encode (body), this.encode (this.secret), 'sha512');
-            headers = {
-                'Key': this.apiKey,
-                'Sign': signature,
-                'Content-Type': 'application/x-www-form-urlencoded',
-            };
-        }
-        return { 'url': url, 'method': method, 'body': body, 'headers': headers };
-    }
-
-    async request (path, api = 'public', method = 'GET', params = {}, headers = undefined, body = undefined) {
-        let response = await this.fetch2 (path, api, method, params, headers, body);
-        if ('result' in response)
-            if (response['result'] !== 'true')
-                throw new ExchangeError (this.id + ' ' + this.json (response));
-        return response;
-    }
-};
